Add ability to insert code prefetches into the CFG. (#11)
Code prefetch instructions are specified as a separate code block. The
code block body is modified to contain either a block of C/asm or a
prefetch instruction specification, indicating which target to prefetch.

Added this capability into the inst_pointer_chase_gen benchmark as a
starting point. We can add this to dfs_chase_gen at a later point.

Issue #10.