Add code prefetch support for DFS chase generator. (#23)
This example is a bit more complex than the instruction pointer chase
CFG. Since we don't know for sure which way the branch will go, we can
either inject a prefetch for both targets at the very top, or we can
duplicate the function body for each side of the branch and prefetch
at the top of the branch. This second option would essentially be
equivalent to the instruction pointer chase CFG, so I've gone with the
first option. This can also be used as a model of uncertainty in the
control flow when injecting prefetches.

Both direct and indirect calls are supported, although ISA support
will vary.

Also, this fixes an issue in which the indirect call mode left out the
useless code block body.